requirements.txt
selenium==3.141.0
newspaper3k==0.2.8
pyperclip==1.7.0
wykop_sdk==0.5.0
pytz==2019.3
